Move ImageComponent to components (FE, #459)
