+++ frontend/src/components/ImageComponent.js
